// src/types/errorMessages.ts
export enum ErrorMessages {
	INVALID_EMAIL = 'Por favor, ingresa un correo electrónico válido',
	INVALID_NAME = 'El nombre no puede tener espacios o caracteres especiales',
	SHORT_PASSWORD = 'La contraseña debe tener más de 5 caracteres',
	LONG_PASSWORD = 'La contraseña debe tener menos de 20 caracteres',
	PASSWORD_DOESNT_MATCH = 'Las contraseñas no coinciden. Por favor, verifica que has escrito las contraseñas correctamente',
	WEAK_PASSWORD = 'Por favor, crea una contraseña que tenga al menos 5 caracteres, 1 carácter en mayúscula, 1 en miúscula y un número',
	FIELD_REQUIRED = 'Este campo es requerido',
	LONG_NAME = 'El nombre debe tener menos de 20 caracteres',
	INVALID_PHONE = 'El número de teléfono no es válido',
}